move listeners to bundles to improve SOC
